feat: Silo Android phone + TV clients
Kotlin Multiplatform phone (com.continuum.app) and Android TV
(com.continuum.app.tv) clients for the Silo self-hosted media server:
Media3 playback with adaptive transcode + capability detection,
library/browse/search, ebook reader, audiobook player, downloads and
offline playback, Watch Together, multi-server and household profiles,
notifications, and admin.

Development history squashed into a single commit.

+++ shared/src/commonMain/kotlin/com/continuum/app/network/api/PersonalDataApi.kt
@@ -99,7 +99,7 @@ class PersonalDataApi(private val client: HttpClient) {
         client.get("/api/v1/ratings/$itemId")
     }
 
-    suspend fun setRating(itemId: String, rating: Double): ApiResult<Unit> = safeApiCall {
+    suspend fun setRating(itemId: String, rating: Int): ApiResult<Unit> = safeApiCall {
         client.put("/api/v1/ratings/$itemId") {
             contentType(ContentType.Application.Json)
             setBody(SetRatingRequest(rating))
